firebase_server/requirements.txt
firebase-admin
python-dotenv
mcp
uv